pyproject.toml
[tool.poetry]
name = "Formidable-Grid"
version = "0.1.0"
description = "Description"
authors = ["Mohammad Rayati"]
readme = "README.md"
package-mode = false 

[tool.poetry.dependencies]
python = "^3.11,<3.14"
flask = "^3.1.2"
itsdangerous = "^2.2.0"
jinja2 = "^3.1.6"
markupsafe = "^3.0.2"
numpy = "^2.3.2"
pandas = "^2.3.2"
certifi = "^2025.8.3"
python-dateutil = "^2.9.0.post0"
pytz = "^2025.2"
six = "^1.17.0"
werkzeug = "^3.1.3"
gunicorn = "^23.0.0"
black = "^25.1.0"
requests = "^2.32.5"
python-dotenv = "^1.1.1"
opendssdirect-py = {extras = ["extras"], version = "^0.9.4"}
plotly = "^6.3.0"
tqdm = "^4.67.1"
ray = "^2.49.1"
psutil = "^7.0.0"
pre-commit = "^4.3.0"
isort = "^6.0.1"
pydantic = "^2.11.9"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
mypy = "^1.16.1"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
internal = []

[tool.isort]
profile = "black"
line_length = 88
combine_as_imports = true
multi_line_output = 3
include_trailing_comma = true
force_sort_within_sections = true
sections = ["STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
known_first_party = ["app", "examples", "tests", "common"]
# force_alphabetical_sort = true
# order_by_type = true
